+++ plone/app/dexterity/__init__.py
@@ -6,10 +6,3 @@ import warnings
 
 _ = ZMessageFactory('plone')
 
-
-def MessageFactory(*args, **kwargs):
-    # BBB Remove in Plone 5.2
-    warnings.warn(
-        "Name clash, now use '_' as usual. Will be removed in Plone 5.2",
-        DeprecationWarning)
-    return _(*args, **kwargs)